Use CGFloat as original structures do and double for scale instead of CGFloat -> double -> float
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@40006 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSView.m
@@ -4238,9 +4238,9 @@ static NSView* findByTag(NSView *view, NSInteger aTag, NSUInteger *level)
   NSPrintInfo *info = [printOp printInfo];
   NSGraphicsContext *ctxt = [printOp context];
   NSSize paperSize = [info paperSize];
-  float scale = (double)([self frame].size.width/paperSize.width);
-  float xoff = 0.0;
-  float yoff = (paperSize.height - [self frame].size.height) / scale;
+  double scale = (double)([self frame].size.width/paperSize.width);
+  CGFloat xoff = 0.0;
+  CGFloat yoff = (paperSize.height - [self frame].size.height) / scale;
 
   [ctxt  beginPage: ordinalNum
          label: aString
